Add saturating addition/subtraction. (#189)
This implements saturating addition/subtraction with inferred widths.

The problem is that you need to know the width of a number to be able to
implement saturating add/sub. If the width is inferred, the frontend
can't know what the max and min are.

To deal with this, we implement a pass that inserts a dummy operation
with the right width inference behavior. This dummy operation is used in
by FIRRTL for width inference, and then a later pass comes in and
removes the dummy operation, replacing it with a saturating add/sub
implementation.